backend/requirements.txt
# FastAPI et serveur
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Base de données
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-decouple==3.8
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# Email
fastapi-mail==1.4.1

# Storage (Supabase)
supabase==2.11.0

# Validation et data
email-validator==2.2.0
phonenumbers==8.13.50

# CORS et middleware
fastapi-cors==0.0.6

# Utilitaires
python-dateutil==2.9.0
pytz==2024.2

# Cache (optionnel)
redis==5.0.8

# Développement
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
faker==33.1.0
requests==2.32.3
